fix : 네비게이션 UI 도감, 랭킹 추가
