+++ src/components/NavigationBar/styles.module.scss
@@ -2,18 +2,28 @@
   position: fixed;
   top: 0;
   width: 100%;
-  height: 85px;
-  align-items: center;
-  background-color: var(--background-color);
-  z-index: 100;
-}
-
-.searchSection {
-  display: flex;
+  max-width: 1400px;
+  height: 5rem;
   align-items: center;
   justify-content: space-between;
-  width: 90vw;
+  background-color: var(--background-color);
+  z-index: 100;
+  background: rgba(255, 255, 255, 0.15);
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  backdrop-filter: blur(20px);
+}
+
+.subcontainer {
+  display: flex;
+  align-items: center;
   max-width: 1400px;
+}
+.searchContainer {
+  display: flex;
+  align-items: center;
+  margin-right: 1rem;
+  width: 15rem;
+  height: 2rem;
 }
 
 .title {
@@ -33,18 +43,27 @@
 }
 
 .logoName {
-  display: flex;
-  align-items: center;
-  font-size: 2.5rem;
+  font-size: 1.6rem;
   font-family: "Sunshiney", cursive;
 }
-
-.searchContainer {
-  display: flex;
-  align-items: center;
-  margin-right: 1rem;
-  width: 15rem;
-  height: 2rem;
+.logosubName {
+  font-size: 0.9rem;
+}
+.encyBTN {
+  color: #3d8a71;
+  margin-left: 3rem;
+  font-size: 2rem;
+  &:hover {
+    color: #88c8b3;
+  }
+}
+.rankBTN {
+  color: #3d8a71;
+  margin-left: 2rem;
+  font-size: 2rem;
+  &:hover {
+    color: #88c8b3;
+  }
 }
 
 .searchInput {
@@ -85,16 +104,22 @@
   }
 }
 
-.navline {
-  width: 90vw;
-  height: 5px;
-  background-color: white;
-  border-radius: 10px;
-  max-width: 1400px;
-  margin-left: 1rem;
-  margin-right: 1rem;
-}
-
 .searchIcon {
   width: 1.5rem;
+}
+
+.rankOverlay {
+  position: fixed;
+  display: flex;
+  justify-content: space-evenly;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.8);
+  z-index: 200;
+  animation-duration: 0.25s;
+  animation-timing-function: ease-out;
+  animation-name: fadeIn;
+  animation-fill-mode: forwards;
 }
